Class-3/Tensorflow/fruits_recognition_tf2/requirements.txt
tensorflow == 2.10.1
pyqt5
pillow
opencv-python